Reorganize & refactor, and improve tests.
This splits up the test & script into multiple utility packages (for
common & testing utilities), reorganizes the test to be in a proper
javatests/ & java/ arrangement, and adds tests for checking the error
conditions of the test.
