+++ scripts/BUILD.bazel
@@ -4,11 +4,33 @@ Kotlin-based scripts to help developers or perform continuous integration tasks.
 
 load("@io_bazel_rules_kotlin//kotlin:kotlin.bzl", "kt_jvm_binary")
 
-# Visibility for libraries that have binaries defined at this package.
+# Visibility for libraries that should be accessible to script tests.
 package_group(
-    name = "oppia_script_lib_visibility",
+    name = "oppia_script_test_visibility",
+    packages = [
+        "//scripts/src/javatests/...",
+    ],
+)
+
+# Visibility for libraries that have binaries defined at this package & tests.
+package_group(
+    name = "oppia_script_binary_visibility",
+    includes = [
+        ":oppia_script_test_visibility",
+    ],
     packages = [
         "//scripts",
+    ],
+)
+
+# Visibility for libraries that should be accessible to other script packages & tests.
+package_group(
+    name = "oppia_script_library_visibility",
+    includes = [
+        ":oppia_script_test_visibility",
+    ],
+    packages = [
+        "//scripts/src/java/...",
     ],
 )
