+++ scripts/src/javatests/org/oppia/android/scripts/ci/BUILD.bazel
@@ -1,0 +1,18 @@
+"""
+Tests corresponding to to developer scripts that help with continuous integration workflows.
+"""
+
+load("@io_bazel_rules_kotlin//kotlin:kotlin.bzl", "kt_jvm_test")
+
+kt_jvm_test(
+    name = "ComputeAffectedTestsTest",
+    srcs = ["ComputeAffectedTestsTest.kt"],
+    deps = [
+        "//scripts/src/java/org/oppia/android/scripts/ci:compute_affected_tests_lib",
+        "//scripts/src/java/org/oppia/android/scripts/testing:test_bazel_workspace",
+        "//scripts/src/java/org/oppia/android/scripts/testing:test_git_workspace",
+        "//testing:assertion_helpers",
+        "//third_party:com_google_truth_truth",
+        "//third_party:org_jetbrains_kotlin_kotlin-test-junit",
+    ],
+)
